+++ GraphicsPipelineIntegratedBRDF.cpp
@@ -29,7 +29,7 @@ void GraphicsPipelineIntegratedBRDF::createDescriptorPool(LogicalDevice& logical
     poolSizes[0].descriptorCount = 1;
 
     poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
-    poolInfo.poolSizeCount = poolSizes.size();
+    poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
     poolInfo.pPoolSizes = poolSizes.data();
     poolInfo.maxSets = 1;
 
@@ -52,7 +52,7 @@ void GraphicsPipelineIntegratedBRDF::createDescriptorSetLayout(LogicalDevice& lo
     bindings[0].stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
 
     layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-    layoutInfo.bindingCount = bindings.size();
+    layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
     layoutInfo.pBindings = bindings.data();
 
     if (VK_SUCCESS != vkCreateDescriptorSetLayout(logicalDevice.getDevice(), &layoutInfo, nullptr, &descriptorSetLayout))
